requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
anthropic>=0.25.0
pdfplumber==0.11.0
openpyxl==3.1.2
python-docx==1.1.2
Pillow==10.3.0
sqlalchemy==2.0.30
